+++ Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts/Definitions/AccountStatus.cs
@@ -1,15 +1,16 @@
-﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions;
-
-/// <summary>
-/// Campaign status filters.
-/// </summary>
-public enum AccountStatus
+﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions
 {
+    /// <summary>
+    /// Campaign status filters.
+    /// </summary>
+    public enum AccountStatus
+    {
 #pragma warning disable CS1591 // self-explanatory
-    ACTIVE,
-    DRAFT,
-    CANCELED,
-    PENDING_DELETION,
-    REMOVED,
+        ACTIVE,
+        DRAFT,
+        CANCELED,
+        PENDING_DELETION,
+        REMOVED,
 #pragma warning restore CS1591 // self-explanatory
+    }
 }
