+++ Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts/Definitions/AccountType.cs
@@ -1,12 +1,13 @@
-﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions;
-
-/// <summary>
-/// Campaign type filters.
-/// </summary>
-public enum AccountType
+﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions
 {
+    /// <summary>
+    /// Campaign type filters.
+    /// </summary>
+    public enum AccountType
+    {
 #pragma warning disable CS1591 // self-explanatory
-    BUSINESS,
-    ENTERPRISE,
+        BUSINESS,
+        ENTERPRISE,
 #pragma warning restore CS1591 // self-explanatory
+    }
 }
