+++ Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts/Definitions/IHttpClientFactory.cs
@@ -1,16 +1,17 @@
-﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions;
-
-using System.Net.Http;
-
-/// <summary>
-/// Http Client Factory Interface.
-/// </summary>
-public interface IHttpClientFactory
+﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions
 {
+    using System.Net.Http;
+
     /// <summary>
-    /// Create client
+    /// Http Client Factory Interface.
     /// </summary>
-    /// <param name="options">Option parameters</param>
-    /// <returns>HttpClient object</returns>
-    HttpClient CreateClient(Options options);
+    public interface IHttpClientFactory
+    {
+        /// <summary>
+        /// Create client
+        /// </summary>
+        /// <param name="options">Option parameters</param>
+        /// <returns>HttpClient object</returns>
+        HttpClient CreateClient(Options options);
+    }
 }
