+++ Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts/Definitions/StatusFilter.cs
@@ -1,13 +1,14 @@
-﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions;
-
-/// <summary>
-/// Filter object used with LinkedIn request.
-/// </summary>
-public class StatusFilter
+﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions
 {
     /// <summary>
-    /// Account status filter.
+    /// Filter object used with LinkedIn request.
     /// </summary>
-    /// <example>AccountStatus.ACTIVE</example>
-    public AccountStatus AccountStatus { get; set; }
+    public class StatusFilter
+    {
+        /// <summary>
+        /// Account status filter.
+        /// </summary>
+        /// <example>AccountStatus.ACTIVE</example>
+        public AccountStatus AccountStatus { get; set; }
+    }
 }
