+++ Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts/Definitions/TypeFilter.cs
@@ -1,13 +1,14 @@
-﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions;
-
-/// <summary>
-/// Filter object used with LinkedIn request.
-/// </summary>
-public class TypeFilter
+﻿namespace Frends.LinkedIn.SearchAdAccounts.Definitions
 {
     /// <summary>
-    /// Account type filter.
+    /// Filter object used with LinkedIn request.
     /// </summary>
-    /// <example>AccountType.BUSINESS</example>
-    public AccountType AccountType { get; set; }
+    public class TypeFilter
+    {
+        /// <summary>
+        /// Account type filter.
+        /// </summary>
+        /// <example>AccountType.BUSINESS</example>
+        public AccountType AccountType { get; set; }
+    }
 }
