+++ Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts/Frends.LinkedIn.SearchAdAccounts.csproj
@@ -1,9 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <TargetFramework>net6.0</TargetFramework>
+	  <TargetFrameworks>net6.0;netstandard2.0;net471</TargetFrameworks>
 	  <LangVersion>Latest</LangVersion>
-	  <Version>1.0.0</Version>
+	  <Version>1.0.1</Version>
 	  <Authors>Frends</Authors>
 	  <Copyright>Frends</Copyright>
 	  <Company>Frends</Company>
@@ -15,6 +15,10 @@
 	  <PackageProjectUrl>https://frends.com/</PackageProjectUrl>
 	  <RepositoryUrl>https://github.com/FrendsPlatform/Frends.LinkedIn/tree/main/Frends.LinkedIn.SearchAdAccounts</RepositoryUrl>
   </PropertyGroup>
+
+	<ItemGroup Condition="'$(TargetFramework)'=='net471'">
+		<Reference Include="System.Net.Http" />
+	</ItemGroup>
 
   <ItemGroup>
     <None Include="FrendsTaskMetadata.json" Pack="true" PackagePath="/">
